Move paginated classes to dto
