+++ src/api/dto/ContactDto.ts
@@ -17,8 +17,7 @@ import {
   ValidateNested
 } from "class-validator";
 
-import { GetPaginatedQuery } from "@api/data/PaginatedData/interface";
-
+import { GetPaginatedQuery } from "@api/dto/BaseDto";
 import {
   GetContactProfileDto,
   UpdateContactProfileDto
